backend/pyproject.toml
[tool.poetry]
name = "sdo-mm"
version = "0.1.0"
description = ""
authors = ["Alex Pavlov <yukiu217@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.8.1"
attrs = "25.3.0"
certifi = "2025.1.31"
charset-normalizer = "3.4.1"
django = "5.1.6"
django-cors-headers = "4.7.0"
django-filter = "25.1"
django-ipware = "7.0.1"
django-phonenumber-field = "8.0.0"
django-rest-authemail = "2.1.7"
djangorestframework = "3.15.2"
drf-spectacular = "0.28.0"
drf-yasg = "1.21.9"
idna = "3.10"
inflection = "0.5.1"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
packaging = "24.2"
phonenumbers = "9.0.0"
pillow = "11.1.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "1.0.1"
python-ipware = "3.0.0"
pytz = "2025.1"
pyyaml = "6.0.2"
referencing = "0.36.2"
requests = "2.32.3"
rpds-py = "0.23.1"
sqlparse = "0.5.3"
typing-extensions = "4.12.2"
tzdata = "2025.1"
uritemplate = "4.1.1"
urllib3 = "2.3.0"
gunicorn = "^23.0.0"
loguru = "^0.7.3"
pydantic = "^2.10.6"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
